[KK-39] изменил цвета на AppColors
