+++ kurs.kz/Offer/OfferTableViewHeaderView.swift
@@ -23,15 +23,14 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
         let label = UILabel()
         label.text = "Заявка"
         label.font = .systemFont(ofSize: 16, weight: .semibold)
-        label.textColor = UIColor(named: "OrderLabel")
+        label.textColor = AppColor.gray100.uiColor
         return label
     }()
 
     private let cancelButton: UIButton = {
         let button = UIButton(type: .system)
         button.setTitle("Отменить", for: .normal)
-        button.tintColor = .gray
-        button.backgroundColor = .white
+        button.tintColor = AppColor.gray80.uiColor
         button.titleLabel?.font = UIFont.systemFont(ofSize: 14, weight: .medium)
         return button
     }()
@@ -45,17 +44,17 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
         textField.selectedTitleColor = AppColor.gray50.uiColor
 
         textField.placeholderFont = UIFont.systemFont(ofSize: 16.0)
-        textField.placeholderColor = UIColor(named: "phoneNumberColor")!
+        textField.placeholderColor = AppColor.gray100.uiColor
         textField.placeholder = "5 000 $"
 
-        textField.backgroundColor = .systemGray6
+        textField.backgroundColor = AppColor.gray10.uiColor
         textField.lineView.isHidden = true
         textField.font = UIFont.systemFont(ofSize: 16.0)
-        textField.layer.borderColor = UIColor(named: "phoneTextFieldBorderColor")?.cgColor
+        textField.layer.borderColor = AppColor.gray20.uiColor.cgColor
         textField.layer.borderWidth = 1
         textField.layer.cornerRadius = 8
 
-        textField.keyboardType = .phonePad
+        textField.keyboardType = .numberPad
         textField.setTitleVisible(true)
         return textField
     }()
@@ -69,23 +68,23 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
         textField.selectedTitleColor = AppColor.gray50.uiColor
 
         textField.placeholderFont = UIFont.systemFont(ofSize: 16.0)
-        textField.placeholderColor = UIColor(named: "phoneNumberColor")!
+        textField.placeholderColor = AppColor.gray100.uiColor
         textField.placeholder = "250 000 т"
 
-        textField.backgroundColor = .systemGray6
+        textField.backgroundColor = AppColor.gray10.uiColor
         textField.lineView.isHidden = true
         textField.font = UIFont.systemFont(ofSize: 16.0)
-        textField.layer.borderColor = UIColor(named: "phoneTextFieldBorderColor")?.cgColor
+        textField.layer.borderColor = AppColor.gray20.uiColor.cgColor
         textField.layer.borderWidth = 1
         textField.layer.cornerRadius = 8
 
-        textField.keyboardType = .phonePad
+        textField.keyboardType = .numberPad
         textField.setTitleVisible(true)
         return textField
     }()
 
-    private let maintextField: NurbolTextField = {
-        let textField =  NurbolTextField()
+    private let maintextField: SkyFloatingLabelTextField = {
+        let textField =  OfferSkyFloatingLabelTextField()
         textField.titleFormatter = {$0}
         textField.titleLabel.font = UIFont.systemFont(ofSize: 12.0)
         textField.title = "По курсу"
@@ -93,16 +92,16 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
         textField.selectedTitleColor = AppColor.gray50.uiColor
 
         textField.placeholderFont = UIFont.systemFont(ofSize: 16.0)
-        textField.placeholderColor = UIColor(named: "phoneNumberColor")!
+        textField.placeholderColor = AppColor.gray100.uiColor
         textField.placeholder = "500 т"
 
-        textField.backgroundColor = .systemGray6
+        textField.backgroundColor = AppColor.gray10.uiColor
         textField.lineView.isHidden = true
         textField.font = UIFont.systemFont(ofSize: 16.0)
-        textField.layer.borderColor = UIColor(named: "phoneTextFieldBorderColor")?.cgColor
+        textField.layer.borderColor = AppColor.gray20.uiColor.cgColor
         textField.layer.borderWidth = 1
         textField.layer.cornerRadius = 8
-        textField.keyboardType = .phonePad
+        textField.keyboardType = .numberPad
         textField.setTitleVisible(true)
         return textField
     }()
@@ -110,7 +109,7 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
     private let changeButton: UIButton = {
         let button = UIButton(type: .system)
         button.setTitle("Изменить", for: .normal)
-        button.tintColor = .systemBlue
+        button.tintColor = AppColor.primaryBase.uiColor
         button.titleLabel?.font = UIFont.systemFont(ofSize: 14, weight: .medium)
         return button
     }()
@@ -138,6 +137,7 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
             containerView.addSubview($0)
         }
         contentView.addSubview(containerView)
+        contentView.backgroundColor =  AppColor.gray10.uiColor
     }
     // MARK: - Setup Constraints
 
@@ -176,46 +176,5 @@ class OfferTableViewHeaderView: UITableViewHeaderFooterView {
             make.trailing.equalToSuperview().offset(-16)
             make.height.equalTo(56)
         }
-    }
-}
-
-final class NurbolTextField: SkyFloatingLabelTextField {
-    override func rightViewRect(forBounds bounds: CGRect) -> CGRect {
-        let padding: CGFloat = 16.0
-        let rightViewWidth: CGFloat = 24.0
-        let rightViewHeight: CGFloat = 24.0
-        let middlePointAdjuster = (bounds.height - rightViewHeight) / 2.0
-        let rightBounds = CGRect(x: bounds.width - rightViewWidth - padding,
-                                 y: middlePointAdjuster,
-                                 width: rightViewWidth,
-                                 height: rightViewHeight)
-        return rightBounds
-    }
-
-    override func editingRect(forBounds bounds: CGRect) -> CGRect {
-        let trimValue: CGFloat = 30
-        let rightViewWidth: CGFloat = 40
-        return CGRect(x: 10, y: 0, width: bounds.width - rightViewWidth - trimValue, height: bounds.height)
-    }
-
-    let padding = UIEdgeInsets(top: 10, left: 16, bottom: 0, right: 0)
-
-    override public func textRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: padding)
-    }
-
-    override public func placeholderRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: padding)
-    }
-//
-//    override public func editingRect(forBounds bounds: CGRect) -> CGRect {
-//        return bounds.inset(by: padding)
-//    }
-
-    override public func titleLabelRectForBounds(_ bounds: CGRect, editing: Bool) -> CGRect {
-        if editing {
-            return CGRect(x: 16, y: 8, width: bounds.size.width, height: titleHeight())
-        }
-        return CGRect(x: 100, y: titleHeight(), width: bounds.size.width, height: titleHeight())
     }
 }
