+++ kurs.kz/SignUp/CustomSkyFloatingLabelTextField.swift
@@ -10,7 +10,7 @@ import SkyFloatingLabelTextField
 
 final class CustomSkyFloatingLabelTextField: SkyFloatingLabelTextField {
     
-    let padding = UIEdgeInsets(top: 10, left: 16, bottom: 0, right: 0)
+    let padding = UIEdgeInsets(top: 10, left: 16, bottom: 0, right: 16)
 
     override public func textRect(forBounds bounds: CGRect) -> CGRect {
         return bounds.inset(by: padding)
